writeup/sploits/requirements.txt
websocket-client==1.8.0
requests==2.32.4
bcrypt==4.3.0